stream: stream_recv writes data to a provided buffer
The buffer is provided from 'c' as a pointer to an array of uint8_t and
a length thereof.

In addition, from 'c', a reference to a bool is provided so that the fin
status of the read can be returned to the caller.

The return value for the 'c' function is the number of bytes written
into the provided buffer.

The rust interface is similar, in that it also provides a buffer, in
the form of a mutable slice of 'u8' type.  The slice self reports its
own size.

The number of bytes written to the slice, and the fin status of the
read is returned as a tuple.

The important fix is in the RecvBuf class, where the pop method now
takes a slice as a parameter and copies queued pages of data into the
slice.

The existing RangeBuf interface is removed from RecvBuf::pop and
Stream::recv_pop and as a result, RecvBuf no longer needs to be
exposed through the ffi interface.