h3: refactor stream buffering and state machine
This refactors how data is read from transport streams and buffered
at the HTTP/3 layer, as well as simplifies the HTTP/3 stream state
machine.

Instead of reading and buffering a fixed amount of bytes from readable
streams, only the amount of bytes required to complete the current state
is read from the transport. Data is still buffered in the HTTP/3
stream's state in order to deal with partial reads, but is now read
directly into the stream buffer to avoid one intermediate copy.

The stream's buffer is resized accordingly to the number of bytes
required by the next state. However the allocated memory for the buffer
is reused across states to avoid unnecessary allocations (an allocation
is only required if the new state requires more data than the current
state).

As for the state machine, all the states responsible for reading a
varint's length have been merged with the states actually reading the
varint itself, to simplify the state machine.