schedule writable streams fairly
Instead of always picking the first writable stream from the streams map
(depending on the HashMap iteration order, so not necessarily the stream
with the lower ID), keep track of writable streams in a separate FIFO
queue, and round-robin them.

This preserves the order of stream writes to what the application
decides (i.e. the order of stream_send() calls), and increases stream
concurrency and significantly reduces deviation in response times
between concurrent requests.

For example, running 250 identical HTTP/3 requests concurrently using
h2load shows the following before the change:

  finished in 78.66s, 3.18 req/s, 3.12MB/s
  requests: 250 total, 250 started, 250 done, 250 succeeded, 0 failed, 0 errored, 0 timeout
  status codes: 250 2xx, 0 3xx, 0 4xx, 0 5xx
  traffic: 245.25MB (257165011) total, 4.39KB (4500) headers (space savings 58.14%), 238.42MB (250000000) data
                       min         max         mean         sd        +/- sd
  time for request:   133.99ms      78.66s      26.53s      23.46s    60.80%

and after:

  finished in 76.32s, 3.28 req/s, 3.21MB/s
  requests: 250 total, 250 started, 250 done, 250 succeeded, 0 failed, 0 errored, 0 timeout
  status codes: 250 2xx, 0 3xx, 0 4xx, 0 5xx
  traffic: 245.25MB (257166916) total, 4.39KB (4500) headers (space savings 58.14%), 238.42MB (250000000) data
                       min         max         mean         sd        +/- sd
  time for request:     75.61s      76.31s      76.10s    170.77ms    59.20%

In the future this can be adapted to implement a more sophisticated
stream prioritization scheme.

Incidentally this also shows significant improvement in the http3
benchmarks due to the fact that we don't need to loop over all streams
(including the ones that have no data to send) when deciding which
stream to schedule, but only over the ones that are actually writable:

  http3/128000            time:   [4.1817 ms 4.8528 ms 5.4428 ms]
                          change: [-52.216% -45.968% -38.798%] (p = 0.00 < 0.05)
                          Performance has improved.
  http3/256000            time:   [10.424 ms 10.667 ms 10.949 ms]
                          change: [-49.763% -48.050% -46.369%] (p = 0.00 < 0.05)
                          Performance has improved.
  http3/512000            time:   [50.051 ms 50.258 ms 50.446 ms]
                          change: [-49.695% -44.301% -38.402%] (p = 0.00 < 0.05)
                          Performance has improved.
  http3/1000000           time:   [82.858 ms 83.813 ms 84.938 ms]
                          change: [-50.842% -49.848% -48.879%] (p = 0.00 < 0.05)
                          Performance has improved.

Other benchmarks are unchanged.