add stream_shutdown() API
This allows applications to discard incoming or outgoing stream data. For
example an application might want to ignore incoming HTTP request data
in case of errors. Without this, the application would have to consume
all the incoming data in the stream manually, otherwise the HTTP/3 layer
would keep reporting "Data" events forever.

Currently that's all this API does currently, but in the future it can
be used to trigger sending STOP_SENDING and RESET_STREAM frames for
specific streams.