h3: don't error on QPACK instruction
Before this change, *any* data received on a QPACK encoder/decoder stream
would trigger a conn_close with encoder/decoder error type, which could
be the wrong thing to do in some cases. This change allows data on the
streams but since we don't support QPACK dynamic table yet, we simply
put the stream into a draining state and ignore instructions.

With this change in place we can properly detect stream closure, so I
took the chance to update the relevant test.