stream application data API
This allows applications to store stream-specific data inside quiche, so
as to avoid having to track streams by themselves.

One potential use-case was storing internal HTTP/3 stream state, so we
wouldn't need to keep separate streams maps for both transport and
HTTP/3 however the current structure of the HTTP/3 implementation does
not allow for this (due to cross-references that make borrow checker
sad). Though this can be useful for other applications.

Fixes #112.