h3: poll with closure API
This adds a `poll_with()` method that, instead of returning single
events like `poll()` does, passes each event to an application-provided
closure. Notably this means that all available events are passed to the
application, rather than just the first one like `poll()`.

This allows applications to potentially skip particular events while
still processing others, or stop processing before all events are
issued, rather than always getting the same event over and over until it
is processed.

The poll() method is also re-implemented using poll_with(). The stream
and DATAGRAM thresholds only apply to the poll() method, as an
application can more easily decide to skip particular events when using
poll_with(), so additional options to do that are not required.

For DATAGRAMs, due to the fact that each Datagram event needs to be
associated with a flow ID, we can't simply use the `dgram_recv_peek()`
method provided by the transport connection anymore, so the internal
dgram recv queue is accessed directly to ensure all required events are
issued (instead of just issuing the same event for the first DATAGRAM
over and over).