fix empty STREAM frames
Make sure to flush STREAM frames with fin=true even if the stream has
otherwise no data to send. This requires marking the stream as
flushable when the application calls Connection::stream_send() with an
empty buffer, but fin=true.

In addition, on reception of empty STREAM frames with fin=true, make
sure that the stream is marked as readable, so the application can
receive the end of the stream, but only do it once, so we don't keep
storing empty buffers and notifying the application as that could be
used as a DoS vector.