reject undecryptable packets
We try to ignore invalid incoming packets (e.g. ones that fail
decryption) to prevent an attacker from causing the connection to be
closed by injecting invalid packets (this is unlike TCP, where e.g. RST
packets can be injected due to lack of encryption).

However currently this is not applied consistently, and there are
several cases where invalid and unauthenticated packets would cause the
connection to fail.

In addition, one problem with this design is represented by the fact
that packets that look like Initial packets but are in fact not QUIC
packets at all, could be received and trick the server into allocating
connection state.

This is unavoidable as to process Initial packets connection state is
required. However if no other valid packet is received, the connection
will never be closed, due to the fact that the invalid packets are not
considered an error but would not cause the idle timer to be initialized
either. So the connection would stay open and never timeout.

This could be fixed by adding a check for this special case to the
existing code, however it's probably not worth the effort and the added
complexity, since, as mentioned above, there are other ways for an
attacker to cause the connection to fail by injecting packets.

Because of the above reasons, this makes packet decryption failures to
close the connection. In the future we can revisit this and implement
a proper mechanism to ignore invalid packets, that can be applied to all
cases and also cover the idle timer edge case.