build: configure BoringSSL for fuzzing when fuzzing
BoringSSL has a fuzzing mode that allows us to bypass some checks done
during the handshake. This makes it possible to build a fuzzing corpus
that can complete the handshake.

Unfortunately, checking for "cfg!(fuzzing)" doesn't work in build.rs, so
this defines a "fuzzing" feature that can be checked instead, and which
is only enabled by the quiche-fuzz crate.