examples/http3-client: try resending requests that failed due to stream limits
This makes the client resend requests that failed due to missing stream
credits. Before the client would simply timeout because it couldn't send
all requests, but now it will attempt resending, with the assumption
that the server might send us MAX_STREAMS frames to increase the stream
credits.

This also refactors how the requests are generated, and instead of
repeating the logic for every request, it just creates the request data
(headers and body) once, and reuses the values. This means that for
example we won't read the body file for every requests anymore, but only
once.