Confirm the handshake by Handshake ack.
Currently we drop the handshake key when the handshake
completed (but not in sync yet) and send some 1-RTT data
and received 1-RTT ack.

In addition to this, mark the handshake confirmed when
Handshake level + ACK is received and this clears all
crypto data in sender buffer.

This can happen when 1-RTT ack is lost but Handshake ack
is received, resulting Handshake key is not dropped during
the request.

Found during the interop with ATS