dgram: initial support for DATAGRAM extension
With this change, we introduce support for the DATAGRAM frame extension
specified in draft-ietf-quic-datagram-00 and draft-schinazi-quic-h3-datagram-03.
These are optional features identified by `quic-dgram` and `h3-dgram`
respectively, example use `cargo build --features "quic-dgram h3-dgram"`.

Several API methods have been added to the quiche transport library:

* `set_dgram_frames_supported()` enables the optional QUIC extension for a
  connection.
* `set_dgram_send_max_queue_len()` and `set_dgram_recv_max_queue_len()`
  control internal queue sizes (see below).
* `dgram_max_writable_len()` returns the size of the largest DATAGRAM frame
  payload that can be sent on the connection.
* `dgram_send()` buffers the provided payload data in quiche, a subsequent
  call to `send()` constructs a DATAGRAM frame and QUIC packet.
* `dgram_purge_outgoing()` purges buffered payloads matching a predicate.
* `dgram_recv()` extracts payload data buffered in quiche from a prior call
  to `recv()`.

Internally, a send and receive queue are added to quiche for the purposes of
buffering the payload of DATAGRAM frames. By their specification, these frames
are not flow controlled by QUIC but because of quiche's API design we benefit
some ability to enqueue/dequeue application data independent from the
packetization and network loop. The size of queues is configurable in order
for applications to tailor things to their needs.

Several API methods have been added to the quiche HTTP/3 library, which
internally builds on the transport library:

* `dgram_max_writable_len()` returns the size of the largest DATAGRAM frame
  payload that can be sent on the connection.
* `dgram_send()` buffers the payload for a flow ID.
* `poll_dgram()` processes received DATAGRAMs and generates DatagramEvents.

In order to exercise the new APIs, a new `dgram` binary crate is added to tools.
This provides a simple client and server that speak draft-pardue-quic-siduck-00,
and HTTP/3.

Thanks to Marco Mastropaolo for contributions that took this from PoC code to
something more robust and featureful.

Co-authored-by: Marco Mastropaolo <marco@mastropaolo.com>