h3: add has_body flag to HTTP/3 Header Events
Previously, when `quiche::h3::Connection::poll()`
returned a `Headers` event, there was no way to tell
if it simultaneously closed the indicated stream.
Instead, `poll()` would need to be called again
until the `Finished` event occurred.

With this change a boolean `has_body` flag is returned
with the `Headers` Event. If `has_body` is false, then
the stream has completed and applications can make use
of this fact to for a more optimal work flow.

This changes the Rust API, the Headers event is changed
from a `Vec<Header>` to a struct.

This changes the FFI, a new `quiche_h3_event_header_has_body`
method allows for checking the `has_body` flag of a
received event.