examples: add client option for requests to different URLs
Previously we allows only a single URL value on the CLI meaning
that we restricted `client` to making a single HTTP/0.9
request per conection, and `http3-client` to making multiple requests
for the same resource in a single connection if `-n` was provided.

With this change, we extend the CLI to allow a list of URLs.
`client` and `http3-client` then create one connection to the first
host in that list and issue as many requests as required.

When `http3-client` is invoked with multiple URLs together with
`-n` each URL is requested that many times before moving to the next.