travis: split NGINX into its own build
Currently the "stable Linux" build is the one that takes the most time
to complete, which means that the entire build will have to wait for
that to complete.

By splitting NGINX into its own build we lower the time of the "stable
Linux" build. While this goes over the 5 concurrent builds limit, the
"stable Android" is very short, which means we might be able to complete
both that one and the NGINX one before "stable Linux" (or around the
same time).