bundle an ACK frame every 4 packets
When sending a packet, we generally don't bundle ACK frames unless an
ack was elicited. This means that if a client has a large flow control
window we will not ack the client's acks for a long while. The client
will then keep sending bigger and bigger ACK frames, which end-up being
very expensive to process.

Note however that this will not trigger sending an ACK-only packet, just
bundling an ACK frame when other frames were already going to be sent.