h3: only consume stream IDs after a successful write
Instead of increasing the next available stream ID counters, wait until
data (e.g. HEADERS or some control data) has been actually written to
the stream, this way we avoid burning through a bunch of stream IDs in
case of errors (e.g. stream blocked, or insufficient stream credits).

This also fixes a potential integer overflow when 4 is added to the
current stream ID (the if would only check that the previous ID was
below std::u64::MAX, but the result of the sum could go over the limit),
though it's unlikely anyone can actually hit that limit.