path MTU discovery
This adds initial support for PMTUD. The PMTU starts at 1200 bytes
(minimum QUIC packet size), then 1452 is tried (maximum QUIC packet
size). If the probe fails we keep using 1200 and not attempt any
additional probe.

Due to the fact that the probe might be lost due to a lower path MTU, we
also need to ignore PMTUD probe losses for congestion control purposes,
otherwise we might decrease the cwnd even though there is no actual
congestion on the path.

The probing happens immediately after the handshake is completed (on the
server it is sent after HANDSHAKE_DONE, to avoid delaying the handshake
confirmation), so it might be too aggressive as it means the connection
incurs the cost of sending the additional packet even when the total
number of bytes is very small (e.g. potentially if the whole data would
have taken a single flight to send, it will now take 2).

For these reasons, PMTUD is disabled by default, and can be enabled with
a config flag. In the future we can experiment with smarter strategies.