encode transport params as varints
This adds supports for encoding and decoding transport parameters with
the draft-27 format, as well as the old format, depending on the version
negotiated for the connection.

Due to the fact that parameters change depending on the version, we need
to also re-initialize the TLS state on the client side when that
happens.