don't decrease tx_data when data is lost
Before 56fc60b this was needed because tx_data was increased when
sending STREAM frames, so we had to decrease it on loss to avoid
increasing it twice when retransmitting. But now we only increase
tx_data when data is received by the application so don't need to do
this anymore, and it could cause a violation of the peer's flow control
limit in some cases.