ignore invalid unauthenticated packets
Follow-up to e739f554 and 2e04d446.

The idea is that invalid packets that haven't been authenticated
shouldn't cause a connection to be aborted, as they may be injected by
malicious man-in-the-middle or man-on-the-side attackers.

We were already doing this for undecryptable packets, but there were a
number of cases not covered.