h3: add GOAWAY API
Previously, quiche did some basic GOAWAY frame reception validation
internally but did not expose a method to send GOAWAY, nor an event
to inform apps that a GOAWAY had been received.

On the sending side, this change adds the `send_goaway()` method
that an application can use to send a GOAWAY frame that informs the
peer of the intention to close the connection.

On the receiving side, this change adds the `h3::Event::GoAway`
event which is returned by `poll()`.

The GOAWAY frame contains an identifier field, which affects how an
application should continue using the HTTP/3 connection. If a client
receives the frame, it MUST NOT issue requests with a larger stream
ID than indicated. If a server receives the frame, it MUST NOT
issue a push.

Note that applications, not the quiche library, are responsible for
processing request and response messages. Therefore the application
must ensure that GOAWAYS are issued in a semantically valid way. For
instance, a server that sends GOAWAY for stream ID 100 commits to
processing all requests up to that stream ID, unless it explicitly
cancels them or further reduces the limit. An application that
processes requests on stream IDs beyond the limit will cause
problems that cannot be detected by quiche.