Allow custom roots of trust
Original issue: https://github.com/cloudflare/quiche/issues/418

Allows loading custom trusted certificates to perform certificate
validation with.

Adds a test that this works, which meant creating a cert chain we could
validate. Added a script to generate the certificate data on demand (so
that the process for this is at least marginally documented).