nginx: don't try to send HTTP/3 data after stream is closed
When an HTTP/3 is marked "header-only", the respective QUIC stream is
closed immediately after sending the response headers. This also happens
with 0-length body responses, however we still try to send the body
afterwards.

This prevents any data from being sent after the response stream is
closed, similarly to how the HTTP/2 code handles it.