stream: initial support for prioritization
Featuring fixes by @LPardue.

This implements a very simple scheme for stream prioritization at the
transport layer. Each stream can be prioritized with an urgency level (a
number from 0 to 255 where lower is more important), as well as an
"incremental" flag.

Internally, instead of keeping a single queue for all scheduled streams,
an ordered map of lists is used instead. Each list holds the streams
that are scheduled at a given urgency level. When flushing data, the
non-empty list with the lower urgency level is used first.

Each urgency list is in turn split into two queues, one for
non-incremental streams, that are flushed fully in a FIFO order, and one
for incremental ones that are round-robined. Non-incremental streams at
the same urgency level take priority.

Currently priority is only used when flushing streams to the wire, but
not when streams are reported as readable or writable to the application
(yet).

Re-prioritization is not implemented yet.