stream: initial support for prioritization
This implements a very simple scheme for stream prioritization at the
transport layer. Each stream can be prioritized with a number from 0 to
255 (lower is more important).

Currently priority is only used when flushing streams to the wire, but
not when streams are reported as readable or writable to the application
(yet).

Internally, instead of keeping a single queue for all scheduled streams,
an ordered map of lists is used instead. Each list holds the streams
that are scheduled at a given priority. When flushing data, the
non-empty list with the lower priority is used first. Within the same
list, streams are round-robined (so the default behavior doesn't change).

Re-prioritization is handled by removing the stream from the queue
corresponding to the old priority, and adding it to the queue of the new
pririty. However the implementation is very inefficient, due to the fact
that finding and removing the stream from the old queue requires
scanning the whole list linearly, or shifting all the elements in the
queue in the worst case scenario, or a combination of both in most
cases. So until we have a better implementation (I'm looking into it) we
might want to just ignore re-prioritization.