increase anti-amplification limit earlier
As per the comment, this needs to be done as early as possible, so that
we still increase the limit in case of errors and the like. This is
fine, as we only need to keep track of how many bytes the peer sent to
us, whether valid QUIC bytes or not.

This makes it possible to convert a couple other cases where incoming
packets are ignored to use `drop_pkt_on_err()` for consistency.

After this, we _might_ want to remove
db5e6acd59a4d15eb6f79a68d315a60bd729570e, though not sending
CONNECTION_CLOSE on the first invalid packet received might still be a
good idea (e.g. for DDoS prevention purposes).