don't return Error::Done from recv()
This changes the `recv()` public method to never return `Error::Done` to
applications, due to confusion of its meaning and misuse.

Many applications (including the quiche-apps and quiche examples) have
been interpreting the Done error as a signal to stop reading incoming
packets from a socket. However this is wrong, as applications shouldn't
stop reading incoming packets, as that might cause a deadlock where the
socket is not re-armed by the event loop and even if a connection is
closing, new packets should still be processed as the socket might be
shared among multiple connections.

Instead of `Error::Done`, `recv()` will now return an `OK(...)` value
with the full size of the incoming buffer, to mean that the entire
buffer has been processed (or ignored), which was the intended meaning
of Done.