docker: build images from the local git checkout
Currently the Docker images are built from a subdirectory, which
prevents the Dockerfile from accessing files outside of the directory.

This in turn means that in order to build the quiche-apps, the
Dockerfile needs to clone a whole new copy of quiche via git and build
it from scratch.

This is very inefficient (slowing down CI builds), and it also prevents
building images from a non-master branch for local testing.

This change simply shuffles files around to allow the Docker images to
be built from the pre-existing checkout of the quiche repository.