don't explicitly check for early 1-RTT packets
We are not supposed to process incoming 1-RTT packets before the
handshake is completed, so we explicitly check for this situation
when processing incoming packets.

However since f902b76 this is already done by BoringSSL. Notably, it
won't set the read 1-RTT traffic secrets until after the handshake is
completed (but write secrets will be available earlier in order to
support 0.5 RTT on the server). So we don't need to check for this on
the quiche side as the packet will not be decryptable due to lack of
keys.

Note that the "early_1rtt_packet" test still passes after this change,
but if the BoringSSL update from f902b76 is reverted, then the test will
fail (as expected).