cubic: fix HyStart++ cwnd growth during LSS
During LSS, cwnd didn't grow because we reset congestion
recovery start time and `on_packet_acked()` exit early because
it thinks it's in a recovery mode. But there was no congestion
event happened, so LSS should not be considered as a recovery
and cwnd will increase in LSS.

- LSS doesn't initialize congestion recovery start time
  and use `lss_start_time` instead. `lss_start_time` is initialized
  when LSS started.
- hystart.in_lss() is replaced by lss_start_time().
- `now` is added to recovery.hystart_on_packet_acked().

Note that Reno is not impacted.