Rework API and DatagramQueue
Previously, we used a Vec<u8> to transfer DATAGRAM frame
data across the API boundary.

This change makes the datagram API more consistent with
the rest of quiche by using u8 buffers. It extends the
modifications internally all the way to DatagramQueue,
which started life reusing quiche::stream::RangeBuf
when it didn't need to.