build: fix x86_64 iOS build on Xcode 10.1
When we do `cargo lipo` on Xcode 10.1 (11.2 works fine),
it fails with the following error during BoringSSL build:

```
ld: in libcrypto.a(rdrand-x86_64.S.o), building for iOS Simulator,
but linking in object file built for iOS, for architecture x86_64
```

It looks like we need to have `-target x86_64-apple-ios-simulator`
compiler flag for both C/C++ and asm object not to confuse the linker.
This is not added by cmake so manually added in build.rs.

This hack works both Xcode 10.1 and 11.2 (confirmed by CI).
Fixes #515.