authenticate connection IDs
This implements the transport parameters needed to authenticate the
connection IDs introduced in draft-28.

On the server the original_destination_connection_id transport parameter
(renamed from original_connection_id) is now always sent when draft-28
is negotiated (instead of only sending it when stateless retry is
performed), and makes the client enforce this.

Because of this, the server's transport parameters need to be re-encoded
after the client's Initial packet is received. The encoding of transport
parameters is refactored in a dedicated method to avoid code
duplication.

The storage of the peer's original destination connection id is also
moved to after the incoming Initial packet is authenticated. While the
Initial packets encryption keys can be generated by a man-in-the-middle
(so this doesn't necessarily provide any added security) it's generally
good practice to limit the amount of logic performed before
authentication when possible.

In addtion this also implements sending and validating on both endpoints
of the new initial_source_connection_id parameter when draft-28 is
negotiated, as well as sending from the server and validation on the
client of the retry_source_connection_id parameter.