improve keylog API
When an application enables key logging in the Config object, quiche
will automatically read the SSLKEYLOGFILE variable and create the
corresponding file to log to.

However this is slightly dangerous as the application is not necessarily
in control of the environment variable, so keys might accidentally end
up being logged without the application saying so. And it generally
doesn't seem to be good form for a library to depend on environment
variables.

The current API is also not very flexible, as it only allows
applications to log to files, and doesn't allow enabling/disabling
logging on a per-connection basis.

So, in addition to calling `log_keys()` on the Config object,
applications now also need to call the `set_keylog()` method on each
connection they wish to log keys for, passing the `Writer` object that
will receive the actual data from quiche.