try to retransmit old frames on PTO
Currently when sending a PTO probe we either include new data frames or
a PING frame in the packet to make sure that the probe is ack-eliciting.
However if no new data is availabe, we should try to include old data
that has not been marked as lost yet if there is any, instead of just
sending a PING.

This changes that by iterating over old sent packets (that have not been
acked or lost yet) and copying CRYPTO and STREAM frames into the outgoing
packet when creating PTO probes.

This requires cloning the frames, which is somewhat unfortunate, but it
should only happen when PTO expires, which has exponential backoff so it
shouldn't be possible to use this as a DoS vector.