examples: align retry mechanism with Rust examples (#592)
Previously, when doing stateless retry the C example server would
reuse the dcid for new_scid. Once a client sent back the token
and we validated it, a new connection ID would be generated and
sent back to the client as part of the handshake. The client would
detect an error when validating retry_source_connection_id and
close the connection.

This change aligns the C examples to the Rust examples. A new
connection ID is generated and sent in the retry. Once a client
sends back the toked and we validate it, we use the information
at hand instead of generating anything again.

Fixes #566 and #582.