ignore failure to parse payload length
When the type field in a long header packet is corrupted from Handshake
to Initial, the payload length field is instead interpreted as token
length which leads to failure when we try to actually parse the payload
length.

Instead use the `drop_pkt_on_err()` trick when parsing the payload
length fails.

Fixes #614.