stream: allow RangeBuf to be partially consumed without splitting
When "popping" a RangeBuf that exceeded the maximum capacity of the
output buffer, we would simply split-off the internal buffer up to the
maximum cap, and then push a new RangeBuf back into the send or recv
buffer.

This is inefficient for two reasons:

1. Splitting off a Vec means duplicating the remainder of the buffer
   into a new Vec. Depending on the size of the original buffer this
   could be very expensive, which is why we had a hack to split-off
   buffers received by the application before creating the original
   RangeBuf.

2. Pushing back the new RangeBuf into the send or recv buffer is
   potentially expensive as a binary heap is used.

The new popping implementations allow for a RangeBuf to be partially
consumed at its start, without the need for splitting it off.

The downside is that the "consumed" bytes will only be freed when the
whole RangeBuf is consumed, but that should be fine.