h3: allow passing headers by reference
Currently an application has to create a list of `Header` values, which
need to clone the header's name and value as it is an owned type.

In practice however the list of headers is immediately used by the
`send_request()` and `send_response()` methods, so the headers don't
necessarily need to be owned.

This adds a `NameValue` trait which defines the basic methods for any
"header" type (i.e. exposing name and value as string references), which
is then implemented by the existing owned `Header` type and a new
non-owned `HeaderRef` type which only carries around references.

The existing `Header` type is kept, as it is needed to return headers
from `poll()` and it is also used by examples and apps to pass headers
around, and they need an owned type.

The downside is that any Rust application that needs to call `name()` or
`value()` on a header type, now also needs to explciitly import the
`NameValue` trait, which is mildly annoying.