qpack: convert to lowercase during encoding
Instead of immediately converting header names to lower-case when
creating a `Header`, defer the conversion to the huffman encoding phase
as in many cases we don't actually need to convert. This also makes it
possible to avoid modifying the headers.

The main annoyance of this is that the static table lookups now need to
be case insensitive, otherwise we could potentially miss some of the
headers if the applications passes them as upper-case.

Case-insensitive matching is done by turning the static table into an
actual table, and doing the lookup by iterating over it. In practice
this should be equivalent in terms of performance as I belive pattern
matching also compares cases one by one. This also makes it possible to
share the static table between endoder and decoder, without having to
duplicate it.

In the future we might want to further optimize this, e.g. by using a
pre-built trie, or something like that.