stream: use VecDeque to store SendBuf data
This changes the internal storage of SendBuf to a VecDeque, but it
doesn't change the underlying logic. In the future this will allow us to
change how SendBuf data is emitted to avoid the allocation and
additional copy we currently perform.

In most cases insertions to the SendBuf happen at the end of the buffer,
however, due to the fact that buffers need to be stored in order, in
case of retransmissions they might happen at any index in the vector.

While at it, the SendBuf::push_slice() method is also renamed to
SendBuf::write() to be consistent with RecvBuf.