avoid cloning connection IDs when sending packets
This introduces a new "ConnectionId" type that represents a QUIC
connection ID as either a Vec<u8> (as we do right now) or a &[u8].

The latter form allows us to avoid cloning the connection IDs when
generating outgoing packets. When parsing incming packets we still
need to use a Vec though.

Note that this is a backwards-incompatible API change as it modifies
the public Header type.