lib: limiting a burst of outgoing packets
"QUIC draft-32 7.7" says:
>     Sending multiple packets into the network without any delay between
>     them creates a packet burst that might cause short-term congestion
>     and losses.  Senders MUST either use pacing or limit such bursts.
>     Senders SHOULD limit bursts to the initial congestion window; see
>     Section 7.2.  A sender with knowledge that the network path to the
>     receiver can absorb larger bursts MAY use a higher limit.

When I test quiche in a high bandwidth/low rtt environment such as localhost,
cwnd can grow very quickly and can create a lot of packet sent
when 1) many sender packets are acked on once 2) try to recover many losses.

This PR implements limiting outgoing packet burst as recommended in the
draft. Here we limit total number of outgoing packets in a single send() loop
to the initial congestion window. When send() reaches a burst limit,
now `Error::Again` will be returned, so it's up to application to flush
current packets and check if there is incoming packets, or can continue
to allow bursting.

Also note that I added the limit on the send(). Such limiting burst
can be done in the application itself, but it provides more general
support and all quiche application can use same burst limiting.

Even we have a pacing branch, it's a still limited support for a specific
platform and need an application changes. This is more general approach
for all platforms.