make Config and Connection Send + Sync
This makes both Config and Connection Sync-able (in addition to
Send-able which they were already).

The biggest hurdle is the fact that the BoringSSL SSL_CTX and SSL
structures are not thread-safe, so they need to be wrapped in Mutex.

In the case of Config this probably has no real downside, since an
application is not supposed to keep changing it once the initial
configuration is applied.

For Connection, this might affect performance slightly, but mostly just
during the handshake. However some of Connection's methods (like
`is_established()`) could be called repeatedly by the application even
after the handshake, so we "cache" the return value (since they won't
change after the handshake is complete) so we won't need to go through
the mutex anymore.