test: fix a flappy tests::empty_stream_frame
This test is sometimes flappy and it's because when ack delay
encoded in the ACK frame is higher than 63, so it gets 2 bytes.
Tests are running locally so ack delay expected to be very small,
but this is microseconds granularity. For example ack_delay=82
means 656 microseconds (82 << 3, 3 is default ack_delay_exponent)
and it's still small but can happen.

Workaround is to increase ack_delay_exponent parameter. It will
encode ack delay to lower number and makes the possibility to
become higher than 63 very small. I choose 5 here. For example
656 microseconds in the above example now encoded to 20, using
only one 1 varint byte.

Note that there is a still possibility to get ack delay higher
than 63, but there will be much less chance of test failure.

Fixes #712