ci: avoid running duplicate 'push' and 'pull_request' events
The idea is to run builds on the "pull_request" event only for external
PRs (i.e. PRs created from forks of the repository), so that PRs created
from internal branches won't run all builds twice.

This should reduce the number of concurrent builds run for internal PRs
by half, which should reduce the time required to show the green check.

Unfortunately there doesn't seem to be a way to configure the condition
globally, so each build need to duplicate it.