properly handle incoming RESET_STREAM frames
When a RESET_STREAM frame is received, unless the stream was already
readable for other reason, the application wouldn't know about it.

For example, if one peer sent STREAM frames without setting the fin
flag, then sent a RESET_STREAM frame completing the stream (even without
an error), the other peer would correctly close the receive-side of the
stream, but not notify the application.

This would prevent, for example, issuing HTTP/3 events to the application.

This change causes the stream to be marked as readable when a
RESET_STREAM frame is received, as if a STREAM frame with fin=true was
received. In addition, a `StreamReset` error will be returned by the
`stream_recv()` method when called by the application.

This also changes the behaviour in case a RESET_STREAM frame is
received, but either the application hasn't read the data yet, or the
data hasn't been received at all (i.e. there is a gap between the
current known stream offset, and the final size reported in the
RESET_STREAM frame).

In this case we now drop any already buffered data (so it won't be
returned to the application), and any missing data will also be dropped
on receipt. The application will still be notified by making the stream
readable.

For HTTP/3, a new "Reset" event is added, which propagates the error
code from the RESET_STREAM frame up to the application, without causing
the whole connection to be closed when `poll()` is called.