qlog: Always serialize frames field last
In order to preserve the behavior where multiple frames
can be appended on a stream event, the frames field must
serialized last. If that is not the case, it is not
possible to keep the frames array "open" for appending.

This commit fixes Issue #785, specifically. It uses the
preserve_order feature of the serde json library. This
feature tells the library to serialize all structure
fields in the order that they are inserted.

Co-authored-by: Lucas Pardue <lucaspardue.24.7@gmail.com>