apps: implementing limiting send burst
Similar to #684, but this is only for quiche-server in apps.
Should be able to help stable rtt and cwnd for high speed transfer.

In case of continuing write in next cycle, use `continue_write`
flag. If this is true, call poll() with 0 timeout not to wait
for a new incoming packet.