http3_test: add some negative test cases
Previously, http3_test was focused on successful tests where the server
under test was expected to handle requests and provide complete
responses. If an error occurred the test runner would panic and throw
up to the invoker to deal with.

With this change, instead of panicking, the Result is returned directly
to the invoker. This allows for checking success or failure
explicitly.

Several new test cases have been added that ensure the server can handle
clients that violate rules set out in the HTTP/3 spec.