make FFI API optional
Currently the FFI API is always built, whether quiche is used as part of
a Rust project or not. This can be problematic when multiple versions of
quiche are built as part of the same project, as the FFI API would then
conflict. It also doesn't make sense to expose FFI functions if they are
not going to be used.

This adds a new feature (disabled by default) to enable the FFI API.