introduce a possibility for manual stream flow control by an application
This feature allows an application to slow down the reception of packets from the peer.
Without this, the application needs to buffer packets. And since it cannot stop the library
from sending MAX_DATA frames, the buffer can grow to quite large sizes in case the processing
speed is much lower than the packet reception speed.