send CONNECTION_CLOSE on epoch the peer can decrypt
Currently, during handshake, CONNECTION_CLOSE frames are sent in packets
of the highest epoch level the local endpoint can use. However there is
no guarantee that the peer can actually decrypt packets of that epoch.

Instead, track the highest level the peer has used, and use that level
to send errors, this way the peer is most likely to be able to decrypt
the packet.