modify windowed minmax module to be generic over timing values
Previously, minmax was fully generic over the values it stored but used
Duration and Instant for managing windowed minimums and maximums.
While most windowed mins/maxes use wall clock time, BBR's bottleneck
bitrate estimation uses a virtual timer based on RTT (see
draft-cardwell-iccrg-bbr-congestion-control-00 4.1.1.3). This change
enables the use of the Minmax module for this use case by allowing for
the use of any "timing" value so long as it supports the appropriate
operations.