support for sending 0-RTT packets on the client
This adds support for clients to generate 0-RTT packets. Currently only
STREAM and DATAGRAM frames are allowed, though this is somewhat stricter
than what is allowed in theory.

This requies exporting the peer's transport parameters as part of the
TLS session, so that they can be restored on resumption.

This also changes how epoch/packet type is picked when creating a new
packet. Packet types offer more granularity (e.g. Short vs. ZeroRTT), so
it's better to pick the type first and then derive the epoch, rather
than the other way around.

Closes #683.