fix processing peer transport parameters during early data
Commit 877bace8 added support for buffering undecryptable 0-RTT packets
and processing them when 0-RTT read keys become available.

One notable change was the fact that packets could be processed as part
of the `Connection::send()` method, which does call the `do_handshake()`
method for advancing the handshake state, but it doesn't try to parse
and configure the peer's transport parameters, unlike the
`Connection::recv()` method which tries to parse transport params when
receiving CRYPTO frames

This could cause a server to be unable to reply to buffered 0-RTT data
due to missing flow control credits.

This change moves the parsing and configuration of peer transport params
directly into the `do_handshake()` method, so that those actions are
always executed when the handshake is advanced.