ffi: expose client-side session resumption and 0-RTT API
Motivation:

Quiche added support for session resumption and 0-RTT lately but the API was not exposed via the C api.

This was done in https://github.com/cloudflare/quiche/pull/911 and https://github.com/cloudflare/quiche/pull/914.

Modifications:

Add C API to also be able to make use of the functionality via C.

Result:

Be able to use session resumption and 0-RTT via the C api