do not try to coalesce Initial and Short in the same datagram
For Initial packets we need to make the UDP datagram at least 1200
bytes following the transport spec. However when the last packet
in the datagram is Short and the datagram is not fully filled, we try
to do null padding at the end which causes a decryption failure on
the server because Short packet doesn't have an explicit length.

This fix is try to avoid coalescing Initial and Short packet in the
same datagram. It will cause 2 different datagrams so may be
a little inefficient but it only happens when we send Initial and
a small Short packet (mostly, only in the beginning).