don't update app_limited when no new frame generated
When there is no new data or ack to be sent, send_single()
sets app_limited=false and returns Error::Done. When
next ACK received, it mistakenly increases cwnd even though
current bytes-in-flight is much smaller than cwnd.

The fix is to return early with Done when no packet to be
generated in the client.