h3: close QUIC connection when unable to initialize HTTP/3
Previously, if `quiche::h3::with_transport()` failed to initialize the
HTTP/3 connection for any reason, it would return an error and rely on
the caller to tidy up the QUIC connection. This is inconsistent with how
other parts of quiche's HTTP/3 layer deal with low level issues.

This change ensures the QUIC connection is closed before returning the
reason to the caller.