tls: try to sanitize use of raw pointers for Context and Handshake
The requirements for the Sync trait are somewhat confusing [0], but it
appears that the reason why it is not automatically implemented for
pointers is aliasing (that is, the data referenced by the pointer may be
aliased). In addition, non-thread-safe "interior mutable" types
(including pointers) also shouldn't be Sync (see e.g. Rc vs. Arc).

[0] https://doc.rust-lang.org/std/marker/trait.Sync.html

In the case of tls::Context and tls::Handshake, there is no aliasing,
but they do have "accidental interior mutability", where a non-mutable
reference is used to get a mutable pointer in a number of methods.

This means that explicitly implementing both Send and Sync for
tls::Context and tls::Handshake _should_ be safe, after the "accidental
interior mutability" problem mentioned above is fixed. This is done by
making a number of tls::Context and tls::Handshake methods take a
mutable reference.

A side-effect of all this is that we don't need to wrap tls::Context /
tls::Handshake in Mutex anymore, which was somewhat smelly to begin
with, so it's good to remove that as well.