qlog: align time units to qlog draft 02 style
The qlog crate is primarily based on the qlog draft 01 specification.
This allowed configuration of time units used for logging, either in
milliseconds or microseconds. The qlog crate only supports relative time
logging and defaults to millisecond units in integer form.

The unfortunate side effect of millisecond units is that events occurring
close together get aliased into the same millisecond. Tools like qvis
can then misrepresent things.

In qlog draft 02, the time unit handling changed. It is no longer
configurable, instead it is always milliseconds, represented as floating
point. Previously, an event at relative time 100085 microseconds, would
be logged as "100". With this change it is logged as "100.085".

This change updates the time unit handling to align with draft 02. It
removes time unit configurability and logs event times as floating
point.