qlog: convert numbers to actual numbers
In qlog-01, many of the fields that carry numbers were basically defined
as string in order to accommodate perceived problems with JSON and 64-bit
numbers. QUIC is rife with potentially 64-bit numbers but in practice,
many of these fields will contain values that are perfectly legal for
the duration of a connection. In later versions of qlog, many of the
fields have been changed to actual number types.

This change updates many qlog fields and API parameters to be closer the
qlog specifications today by using numbers e.g. stream IDs are now u64
rather than String . There general several benefits to this, we no
longer need to create owned Strings to pass around, we don't waste
effort creating strings that are immediately serialized again by the
streaming serializer, we save space in the serialized logs because
numbers are no longer surrounded with pointless quotes, and we save
space in the buffered serializer.

This change also builds on the recent update to qlog time units.
Several fields in the metrics_updated event are now represented as f32
which encodes milliseconds with decimal precision.

While working through the commit, I was affected by the overloading of
String for qlog numbers and bytes see
https://quicwg.org/qlog/draft-ietf-quic-qlog-main-schema.html#name-bytes.
In order to capture all the cases where we explicitly want to log bytes,
I added a simple type definition, `Bytes`, for disambiguation. In
future, we might substitute this for a more complex type that can
satisfy the qlog requirements.