recovery: only log qlog metrics when fields change
Previously, when the qlog feature was enabled, quiche would log all
recovery metrics (that it cares to log) any time there was a recovery
event. This is legal but increases duplication in the logs.

With this change, we make the event emission smarter by only generating
them if at least one of the event data fields has changed. In testing,
the file size is reduced by up to 25%, the qvis congestion view
continues to work just fine, and the sequence diagram has improved
readability,