write DATAGRAM frame data directly into the output packet
Similarly to how STREAM and CRYPTO frames are written to the output
packet buffer, write the DATAGRAM frame header and data into the output
packet buffer directly.

This doesn't really make much of a difference in terms of performance,
but it enables a later change to use the crypto scatter seal API to
avoid unnecessary copies of data.