qlog: externalize event times and importance
This commit build on the previous one, we can merge if suitable.

qlog 02 adds the time field as a true part of the new Event object. The
code in the prior commit tried to do some handholding to allow people to
pass in events to the streaming serializer and have it generate the
event time. The downside to this is that users need to construct events
with a "null" time that is updated by the streaming serializer before
being written out. That's awkward.

In this change, we make the default 'add_event()` method focus on writing
out event, externalizing the time-setting to the caller. If they really
want help, they can use `add_event_now()` or `add_event_with_instant()`.

The benefit to this approach is that we can make the quiche library code
more straightforward to read. To support this, we'll store our own qlog
start time and calculate the event time ourselves.

I also realised that the way importance works was a bit sub-optimal. An
application that had configured the importance level to be `Core`, could
waste time building event objects that turn out to be `Base` or `Extra`
and therefore would get immediately swallowed by the serializer. So I've
added a new macro, `qlog_with_type_and_start!` that can conditionally
avoid doing work for events outside the current importance level. It
also presents the stored start time for the macro body to access, which
helps to avoid problems related to the conditional compilation of the
qlog feature.