avoid counting reset streams' unsent data towards flow control limit
Currently the send-side flow control allowance is enforced and reduced
whenever stream data is buffered in the corresponding stream's send
buffer (i.e. when the application calls `stream_send()`). This is to
limit the amount of buffered data to what the connection can actually
take and avoid over-buffering.

However whenever a stream is reset (e.g. `STOP_SENDING` is received, or
`stream_shutdown()` is called), the stream's final size will include all
data that was previously buffered, even if not all of it was actually
sent on the wire yet.

To aovid that, we can make the final size be the offset of the first
unsent byte, and then "claw back" the remaining data from the flow
control allowance when a stream is reset.