cubic: proportional rollback threshold
Currently a fixed 10 packets threshold is used for detecting a small
loss. When cwnd is very high (more than 1000 packets, for example),
this makes the detection less sensitive. New approach is to make a
rollback threshold relative to cwnd (currently 20%) to make it work
with various size of cwnd in real time. Minimum threshold is 2.