qlog: refactor events to events module
As part of qlog adoption into the QUIC WG, it was restructured into 3 documents:

* a generic base schema for logs, trace, and events
* a QUIC schema for transport and recovery events
* an HTTP/3 and QPACK schema for application protocol events

Prior to this change, the document was a bit more monolithic and had a supporting
typescript schema that was also monolithic.

The early qlog crate code use the TypeScript schema as a template for structuring
the code, resulting in one single hard-to-manage file. It is anticipated that
the WG will change the qlog schema format from TypeScript to something else
(CDDL?) in the near future. So there is not much benefit sticking to that
template.

The recent changes to support qlog 02 in this crate have given us a lot more
flexibility to refactor the code. This change continues the flexibility theme,
breaking apart the code monolith in order to support future changes to the spec
and any additional qlog extensions that may get defined for other protocols
(such as HTTP/2).