lib: support for JSON-SEQ serialization format
With this change we add a serialization format parameter to the
set_qlog_with_level() method, which allows applications to control
whether qlog is written in JSON or JSON-SEQ format. set_qlog() remains
unchanged and will continue to log in JSON format.

In support of this change, update the quiche_apps with a new
command-line argument to control the format.