h3: ensure StreamBlocked if send_request() blocked
Previously, if the connection-level flow control was 0 when we attempted
to create a new request stream via send_request() we would return the
result of stream_send() directly. This would cause an Error::Done to
be returned, which is not documented in the API.

Rather than have two different error types for a similar class of error,
with this change we ensure that Error::StreamBlocked is returned
consistently.

Fixes #1082.