stream: add a simple hasher for stream collections
quiche uses several different collections to store stream state. To date
many of these use std::collections{HashMap, HashSet} and we've relied on
the default hashing algorithm. However, QUIC stream IDs are a 64-bit
number space that have some protocol-specific behaviors and checks. We
features and properties that many hashing algorithms provide are not so
useful for QUIC stream IDs.

This commit introduces a simple no-op hasher for collections that store
QUIC stream IDs. It just uses the u64 stream identifier directly.
Migrating collections to use the hasher is left to other commits.