Check payload size before decryption
In `packet::decrypt_pkt()`, when `payload` is to small and
no space in `ciphertext`, it may cause overflow in
`max_out_len` in `crypto::open_with_u64_counter()`, can lead
a crash.

This can happen when the received packet payload is corrupted.
To fix it, an overflow check for `max_out_len` is added.