lib: send STREAM_DATA_BLOCKED frames less often
When applications try to send stream data with `stream_send()`, the
number of bytes actually written might be less than supplied due to
insufficient stream or connection capacity. In these cases, quiche
internally marks the stream as blocked and, when `conn.send()` is next
called the mark causes a STREAM_DATA_BLOCKED frame to be emitted in a
new packet.

Previously, every call to `stream_send()` would cause the stream to be
marked and therefore generate a STREAM_DATA_BLOCKED frame. In cases,
where `conn.send()` is called frequently, the blocked frame might be the
only frame in the packet, which is wasteful.

This change reduces the frequency of STREAM_DATA_BLOCKED frames but
avoiding sending them when the offset of sent stream data is unchanged.
When a stream is first blocked during `stream_send()`, a frame is
emitted and the blocked offset is recorded. Subsequent blocked frames
are only emitted if the application is able to write more data and get
blocked at a new offset.