cc: on_packets_acked() hook
Currently we have on_packet_acked() called by each ACK'ed data
packet and it will be called multiple times if an ACK range
contains many data packets. Sometimes there is a need to process
an ACK itself, not the all data packet ACK'ed this time.

To support it, on_packets_acked() hook is introduced in place
of on_packet_acked(). Only difference is passing a list of Acked
instead of each Acked and the hook need to process all packets
in the Acked list if needed, for example to get the total size
of ACK'ed payload.

reno and cubic is updated to use a new hook.