lib: fix retrans_count again
Previously `Stats.lost` is much higher than `Stats.retrans`.
It is because `retrans_count` increases when there is
some retransmitted bytes in the lost packet list, not counting
per retransmitted range.

Now it counts for each lost data frame. `lost` and `retrans`
may not match but in a simple case (e.g. lost 1 packet,
retransmit the same range) it will be same.