http3_test: add RESET_STREAM testing capability
Previously, http3_test just ignored any Event::ResetStream occurrences
that might happen at the client.

With this change we mark reset streams as complete and store the
received error code in the request object, which tests can access for
post exchange asserts.

We use this new capability to add an optional `drip_delay_reset` test to
httpbin_tests, which can be enabled with the `test_resets` feature flag.