Internally rely on a VecDeque to store CIDs and remove breaking changes
Instead of relying on a `HashMap`, we now store CIDs in a wrapper of
`VecDeque`. This wrapper ensures we always have at least one element
in it and also bounds the number of elements inside it.

The code has also be refined to remove breaking changes, namely the
signature of `source_id()` and `destination_id()` are left untouched.