apps: more robust pseudo-header processing
The quiche HTTP/3 layer delegates the responsibility for HTTP message
field processing to applications. This includes processing of
pseudo-header fields.

With this change, we update quiche-server to be more robust in applying
pseudo-header requirements. Requests that fail to meet the requirements
will see that the request stream was reset with the H3_MESSAGE_ERROR
code.