apps: allow apps to be compiled without sfv feature
When we externalized the Priority field value parsing out of quiche,
we provided an optional utility function behind the the `sfv` feature.

In the apps, only the quiche-server requires the `sfv` feature. And we
started by assuming this is a default features. Unfortunately, if trying
to build the quiche-apps crate with `--no-default-features`,
quiche-server will then fail. This is especially annoying because people
have a good reason to build a slimmed-down quiche-client without all the
extra feautres.

This change modifies the quiche-server related code so that is can build
without the `sfv` feature. In this case, any Priority headers or
PRIORITY_UPDATE fields cannot be parsed. Instead, the default priority
is assumed.