avoid storing Connection pointer in TLS ex data
Follow-up to 3c7a597a3765995a5434eb08f838caba65112a6f.

Instead of storing a pointer to the whole `Connection` structure in the
TLS ex data, this change introduces a new `ExData` structure that
provides references to specific `Connection` fields needed in the TLS
callbacks.

This way we avoid having aliased mutable references to a `Connection`
when the BoringSSL callbacks are being executed (which is undefined
behaviour).

This needs to make `QUICHE_EX_DATA_INDEX` private again, since the
`ExData` structure is not public and can't be used by applications
(though nothing should be using this right now).

Fixes #1217.