h3: better support 0-length payload frame parsing
In HTTP/3 some frames can legitimately have 0-length payloads, such as SETTINGS.

Previously, the HTTP/3 frame streaming parsing could get confused by these
frames, whereby it would try to read 0 bytes from the transport layer via
stream_recv(). In some cases, quiche could determine the stream is not
readable, because there are no more bytes to read - imagine a an empty
SETTINGS frame that arrived as the last frame on the control stream. In such a
case, quiche would return Error:Done, which would cascade up to the
HTTP/3 layer's frame parsing and cause it to quit early before parsing
the frame. The insidious thing here is that the HTTP/3 layer would
continue trying to read this stream on every call to `poll()`, but it
would always fail as long as the stream was determined to be not
readable (likely because there is nothing more to send in the control
stream).

With this change, we add a check for this scenario inside the HTTP/3
layer. If a frame's payload is 0-bytes, exit early by returning Ok(())
so that the frame can be consumed immediately. This also helps avoid
wasted work trying to read nothing over and over again.