lib: distinguish length field from payload length
Previously, at `recv_single`, the length field of the packet is named as `payload_len`. Although the decoding process correctly treats it as a length field, the `trace!` log and qlog are not.

According to `send_single`, `trace!` set `len` as the length of the actual plaintext payload. But `trace!` at `recv_single` set `len` as the length field of the packet.

Also according to `send_single` and qlog definition at <https://datatracker.ietf.org/doc/html/draft-ietf-quic-qlog-main-schema-00#section-5.1>, `payload_length` in `RawInfo` should be "the byte length of the entity's payload, without headers or trailers". But at `recv_single`, this `payload_length` in `RawInfo` takes the length field of the packet instead.

After the change, the `trace!` log and qlog are consistent with the decoding process. Also, the variable name is corrected.