lib: don't init qlog pkt hdr unilaterally
qlog packet have a mandatory PacketHeader type that contains several
optional fields. When creating packets to send, some of the fields
needed for qlog would upset the borrow checker later on, so previously
we constructed a PacketHeader early to avoid problems. This was
optionaly compiled via the qlog feature flag but there were no runtime
checks to see if the active connection had a qlog streamer configured.
That lead to unilateral construction of the object in quiche builds
that had the feature but didn't use it, leading to wasted effort.

With this change, the PacketHeader is only constructed if a connection
has actually been configured to output qlogs.

I also did a sweep for any other instances of similar unilateral
construction but this seems to be the only one.