refactor Epoch to be an enum instead of usize consts
Currently, epoch values are just `usize` constants over the code.
The handling of the different epochs is done through arrays and
the epoch-specific structure is accessed using the raw value
behind the constant. While this should not happen, it might be
possible to index a value out-of-bound (for instance, when doing
some arithmetics on epoch values).

This commit replaces the plain `usize` typedef of `Epoch` by a
dedicated `enum`, and implements the `Index(Mut)` traits to
minimize the impact on the existing code (i.e., `Epoch` enum can
be used to index the array).