lib: better Immediate Close during handshake
Previously, when a quiche connection was closed using `close()` with
`app` true, the connection might not already have been in a state
where it was safe to emit the application CONNECTION_CLOSE frame. Because
calling `close()` puts the connection into a terminal state where
newly received packets are discarded rather than processed, it is
possible that a connection in this state cannot emit a suitable
CONNECTION_CLOSE. In some circumstances, this could have led to a
connection becoming idle and then silently closed.

Applications could have worked around this by carefully managing the
use of `close()` based on connection states. For example, by
manually converting to a connection level close (`app` false)
or by deferring the close until a later moment.

This change modifies the `close()` method so that it automatically
detects connection state and follows the recommendations in
https://www.rfc-editor.org/rfc/rfc9000.html#section-10.2.3-3. If
`close()` is called with `app` true, and it is not safe to emit,
the application error information is cleared and replaced with
a connection error of type 0x0c and empty reason.