ensure stream_writable() also checks blocked offset
In PR #1170, we added a change to reduce the frequency of
STREAM_DATA_BLOCKED frames when using stream_send. Unfortunately, we
neglected to apply the same logic to the stream_writable() function.

The h3 layer explicitly calls stream_writable() when it can't write
the full amount of data passed to send_body(). stream_writable marks
the stream as blocked, howver, next the DATA frame gets truncated to
match the stream capacity before calling stream_send(). Inside, it
detects that capacity wasn't exceeded, so the stream isn't actually
and it gets unmarked blocked. It takes the *another* round of
send_body() to cause it to mark the stream as blocked via
stream_writable() and then it exits before any stream_send() (because it
literally can't send anything). This causes the first
STREAM_DATA_BLOCKED to be emitted. Any further calls to send_body()
while in this state will also do the same thing. And because
stream_writable() doesn't check if the offset has changed before marking
the stream, it can cause frames to be sent with duplicate information.

This change simply updates stream_writable() to check the max offset.
The result is that all duplication is removed and we'll only emit one
STREAM_DATA_BLOCKED frame per max stream data change that we receive
(just as it works for stream_send).

To test this out is quite straightforward. Start the quiche-server,
have a 1 GB file ready somewhere, use quiche-client to request it.
Before this change, my ad-hoc testing averaged between 2500 and 3000
STREAM_DATA_BLOCKED frames being sent by the server. After the change,
that number falls to between 60 and 70.

Fixes #1384.