Rewrite congestion and recovery code
This is a very large change that replaces most of the congestion
and recovery control implementation with an implementation that
copies google/quiche for the most part. Most importantly cubic and
pacing are copied exactly from google/quiche so we can have a high
confidence in both. This fixes our pacing problems entirely and
also has a much cleaner interface, where every bit is cleanly separated
from the main state.