FFI: Change SendInfo to return in nanoseconds instead of timespec
# Motivation
Currently, we are returning a `timespec` in the `SendInfo` struct from the FFI interface. To construct that `timespec` we are copying over the bytes from a Rust `Instant` to the `timespec` struct. This runs into undefined behaviour since `Instant` is an opaque type and might not have the same layout as `timespec`. We have noticed this being a problem where Quiche is returning us `timespec`s with the nanosecond part being multiple days.

# Modification
This PR changes the way we return the "write at" information by changing it to a "write in" property. This allows wrapping libraries to use their native `Instant` types to get the current time and then add the returned nanoseconds value to it. This also aligns the interface with how we expose timeout information through the FFI layer.

# Result
We are no longer running into undefined behaviour.