apps: store response priority for retry after block
There are occaisons where the quiche-server will get blocked from
sending response headers. We already cater for this in the application
and store the request headers or body in order to retry once the
stream is reported as writable.

However, in the case of blocked headers, the response priority was
not stored and when the stream was unblocked, quiche-server did not
try to send_response_with_priority(). Together this meant that
responses that were blocked when attempting to send headers would
revert to a default priority.

With this change, we store the original priority that was intended
and make sure to use it on each retry attempt.