h3: better estimate QPACK encoded output size
The HTTP/3 and QPACK encoding process relies the HTTP/3 layer creating
an output buffer and passing this to the QPACK layer for it to encode
into. Previously, the calculation of the buffer size used a simple
approach for estimating the correct size of the buffer. In some cases
this estimate could be wrong, for instance when Huffman-encoding a
literal value in the low or extended ASCII ranges. This could lead to an
underestimation of the buffer size, which cause the QPACK encode to fail
leading to an h3::InternalError that would bubble up to an application
attempting to send requests or responses.

With this change, we move the check of a header's static representation
type earlier in the encode process, and use the type to provide a better
estimate of the required buffer sizes. The additional upshot is that
fully-indexed types (where the name and value are both in the static table)
require less storage and we no longer over estimate.