h3: reduce capacity requirement when sending partial body
It can happen that the congestion window doesn't grow enough to
accomodate a very large body buffer, in which case the corresponding
stream is never reported as writable and it hangs.

Instead, just make sure the stream is reported when at least a few bytes
can be sent. This uses the current write's "overhead" value, which is
not necessarily accurate for future writes, but it should be good
enough for our purposes.