fix handshake confirmation logic
Currently the handshake is confirmed, on both client and server, when an
ACK for a 1-RTT packet is received, per RFC 9002, section 4.1.2 (2nd
paragraph).

However we currently have no way to distinguish ACKs for 1-RTT packets
and 0-RTT ones, which means that a client might end up thinking that the
handshake is confirmed too early, and drop the Handshake epoch
preventing the server from completing the handshake, and because this is
just an optional optimization, it _should_ be safe to remove for the
client.

For the server, we need to implement proper handshake confirmation per
RFC 9002, section 4.1.2 (1st paragraph), which just states that
handshake is confirmed once it is complete.

This also fixes a small bug that caused the "connection established"
trace log to be spuriously emitted when resuming a session, and then
again once the handshake is complete (we should just do the latter).