make path and cid maps infallible
No functional change. Changes path and cid maps to return Option
instead of Result with InvalidState. This is more consistent with
the streams map, pushes the error case higher up in the call stack,
and better reflects what is being returned. For cases where you want
to simply check the existence of a path or cid, an Option is more
natural because there are only two variants: Some or None. When a
Result is returned, any error _could_ be returned. While in most
instances, doc comments indicate that the only Error is InvalidState,
this seems like an unnecessary mental burden.